+++ typing/tests/integration/object.rs
@@ -4,7 +4,7 @@ use assert_matches::assert_matches;
 
 use arithmetic_parser::grammars::Parse;
 use arithmetic_typing::{
-    arith::NumArithmetic, error::ErrorKind, Prelude, TupleLen, Type, TypeEnvironment,
+    arith::NumArithmetic, error::ErrorKind, Assertions, Prelude, TupleLen, Type, TypeEnvironment,
 };
 
 use crate::{hash_fn_type, ErrorsExt, F32Grammar};
@@ -406,5 +406,111 @@ fn tuples_with_dyn_length_as_object_fields() {
     assert_eq!(
         type_env["test"].to_string(),
         "for<'T: { xs: [Num], ys: [Num; N] }> ('T) -> Bool"
+    );
+}
+
+#[test]
+fn object_destructure_basics() {
+    let code = r#"
+        { x } = #{ x = 1; };
+        { x -> y } = #{ x = 2; };
+        obj = #{ xs = (3, 4, 5); flag = 1 == 1; };
+        { xs: (head, ...tail), flag } = obj;
+    "#;
+    let block = F32Grammar::parse_statements(code).unwrap();
+    let mut type_env = TypeEnvironment::new();
+    type_env.process_statements(&block).unwrap();
+
+    assert_eq!(type_env["x"], Type::NUM);
+    assert_eq!(type_env["y"], Type::NUM);
+    assert_eq!(type_env["head"], Type::NUM);
+    assert_eq!(type_env["tail"], Type::from((Type::NUM, Type::NUM)));
+    assert_eq!(type_env["flag"], Type::BOOL);
+}
+
+#[test]
+fn object_destructure_in_fn_args() {
+    let code = r#"
+        test = |{ x, y }| x + y;
+        test(#{ x = 1; y = 2; })
+    "#;
+    let block = F32Grammar::parse_statements(code).unwrap();
+    let mut type_env = TypeEnvironment::new();
+    let output = type_env.process_statements(&block).unwrap();
+
+    assert_eq!(output, Type::NUM);
+    assert_eq!(
+        type_env["test"].to_string(),
+        "for<'T: { x: 'U, y: 'U }, 'U: Ops> ('T) -> 'U"
+    );
+}
+
+#[test]
+fn object_destructure_with_complex_bindings() {
+    let code = r#"
+        test = |{ xs: (head, ...xs), ys }| xs == ys && head == 1;
+        test(#{ xs = (1, 2, 3); ys = (2, 4); });
+        test(#{ xs = (1, true); ys = (true,); });
+    "#;
+    let block = F32Grammar::parse_statements(code).unwrap();
+    let mut type_env = TypeEnvironment::new();
+    type_env
+        .insert("true", Type::BOOL)
+        .process_statements(&block)
+        .unwrap();
+
+    assert_eq!(
+        type_env["test"].to_string(),
+        "for<'T: { xs: (Num, ...['U; N]), ys: ['U; N] }> ('T) -> Bool"
+    );
+}
+
+#[test]
+fn object_destructure_in_map_pipeline() {
+    let code = r#"
+        test = |xs| xs.map(|{ x, y }| x as Num + y);
+        (#{ x = 1; y = 2; }, #{ x = 3; y = 4; }).test() == (3, 7);
+        (
+            #{ x = 1; y = 2; z = 3; },
+            #{ x = 3; y = 4; z = x - y; },
+        ).test() == (3, 7);
+        // Unfortunately, having `z` in *one* of tuple items doesn't work for now.
+    "#;
+    let block = F32Grammar::parse_statements(code).unwrap();
+
+    let mut type_env = TypeEnvironment::new();
+    type_env
+        .insert("map", Prelude::Map)
+        .process_statements(&block)
+        .unwrap();
+    assert_eq!(
+        type_env["test"].to_string(),
+        "for<'T: { x: Num, y: Num }> (['T; N]) -> [Num; N]"
+    );
+}
+
+#[test]
+fn object_destructure_in_fold_pipeline() {
+    let code = r#"
+        minmax = |xs| xs.fold(#{ min = INF; max = -INF; }, |{ min, max }, x| #{
+            min = if(x < min, x, min);
+            max = if(x > max, x, max);
+        });
+        assert_eq((5, -4, 6, 9, 1).minmax(), #{ min = -4; max = 9; });
+    "#;
+    let block = F32Grammar::parse_statements(code).unwrap();
+
+    let mut type_env = TypeEnvironment::new();
+    type_env
+        .insert("INF", Type::NUM)
+        .insert("if", Prelude::If)
+        .insert("fold", Prelude::Fold)
+        .insert("assert_eq", Assertions::AssertEq)
+        .process_with_arithmetic(&NumArithmetic::with_comparisons(), &block)
+        .unwrap();
+
+    assert_eq!(
+        type_env["minmax"].to_string(),
+        "([Num; N]) -> { max: Num, min: Num }"
     );
 }